Fix polyfills
